fix(tracing): send correct SDK spans
Before this change, the Haskell Trace SDK did not differentiate
correctly between SDK spans and registered spans. In particular, the
user provided span name was sent as span.n, instead of data.sdk.name.
Furthermore, data.skd.type was missing. Finally, it allowed to add
arbitrary data to non-registered spans instead of constraining it to
data.sdk.custom.tags.

With this change, all functions that start spans now take a SpanType
argument (instead of a Data.Text for the span name) which can either
denote an SDK span or a registered span.

The SdkSpan type is an instance of Data.String.IsString, so that
string literals can still be passed for SpanType arguments, those will
automatically be converted to SdkSpan values (with the
OverloadedStrings language extension).

BREAKING:

Several functions were renamed or have changed their signature.
In detail:

* Use InstanaSDK.addTag instead of InstanaSDK.addData (for SDK
  spans).
* Use InstanaSDK.addTagAt instead of InstanaSDK.addDataAt (for SDK
  spans).
* For registered spans, replace InstanaSDK.addData(At) with
  InstanaSDK.addRegisteredData(At). Note that you should probably not
  create registered spans, but only use SDK spans.
* Usages like startEntry "some.span.name" or withEntry "some.span.name"
  (that is, the span name is passed directly as a literal) will
  simply continue to work as expected but will require
  OverloadedStrings to be active.
* Usages where the span name is stored in a Data.Text value first and
  then passed to startEntry/withEntry/etc. will break. You can fix those
  by importing

    import qualified Instana.SDK.Span.SpanType as SpanType

and then wrapping the span name in SpanType.SdkSpan. For example:

     spanName = T.pack "some.span.name"
     InstanaSDK.withRootEntry instana spanName ...

becomes:

     spanType = SpanType.SdkSpan $ T.pack "some.span.name"
     InstanaSDK.withRootEntry instana spanType ...