refactor(tracing): rename SDK.addAnnotationAt to addJsonValueAt
Also, fix the format of the span.data.http.header annotation. This was
previously a JSON array of JSON arrays (representing each header as
an array with two elements, name and value), which was incorrect. The
correct format is a JSON object, using the header name as the key and
the header value as the value.

Also, add more functions to add annotations to spans:
* addAnnotationAt - now takes an Annotation value instead of arbitrary
  ToJSON types), the previous addAnnotationAt signature is still
  available as addJsonValueAt
* addAnnotationToEntrySpan
* addAnnotationToEntrySpanAt
* addAnnotationValueToEntrySpanAt
* addJsonValueToEntrySpanAt

BREAKING CHANGE: rename SDK.addAnnotationAt to addJsonValueAt