fix(announce): remove obsolete check for Server header
This check actually breaks the announce procedure if a service mesh with
proxies is present, as the proxy might very well be configured to
replace the Instana agent's Server header with its own. The check is
superfluous anyway. Instead, the trace SDK now checks whether the HTTP
status code is in the 2xx range.